Move Notification event to gaphor.event
So every component can send a message to the user if it needs to.
